% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{create_subscription}
\alias{create_subscription}
\title{Create new subscription}
\usage{
create_subscription(project, subscription, topic, pushConfig = NULL,
  ackDeadlineSeconds = 10)
}
\arguments{
\item{project}{Name of Google Cloud Project}

\item{subscription}{Name of the subscription available for consuming messages}

\item{topic}{The name of the topic from which this subscription is receiving messages.}

\item{pushConfig}{Named list of key and value pairs that would be used as attributes}

\item{ackDeadlineSeconds}{This value is the maximum time after a subscriber receives
a message before the subscriber should acknowledge the message. It will only take a minimum of
10s and a maximum of 600s}
}
\description{
Create new subscription
}

