% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{delete_subscriptions}
\alias{delete_subscriptions}
\title{Delete subscriptions}
\usage{
delete_subscriptions(project, subscription)
}
\arguments{
\item{project}{Name of Google Cloud Project}

\item{subscription}{Name of the subscription available for consuming messages}
}
\description{
Delete subscriptions
}

