% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{list_subscriptions}
\alias{list_subscriptions}
\title{List Subscriptions available in current Google Pubsub project}
\usage{
list_subscriptions(project, pageSize, pageToken)
}
\arguments{
\item{project}{Name of Google Cloud Project}

\item{pageSize}{Maximum number of topic names to return.}

\item{pageToken}{The value returned by the last ListTopicsResponse; indicates that this
is a continuation of a prior topics.list call, and that the system should return the
next page of data.}
}
\description{
List Subscriptions available in current Google Pubsub project
}

