% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{publish_message}
\alias{publish_message}
\title{Publish a message to the current Google Pubsub project}
\usage{
publish_message(project, topic, data, attribute_list)
}
\arguments{
\item{project}{Name of Google Cloud Project}

\item{topic}{Name of the topic to publish to}

\item{data}{The message payload. You do not need to encode the message. The function internally will encode
your message}

\item{attribute_list}{Optional attributes for this message. Only accepts a named list}
}
\description{
Publish a message to the current Google Pubsub project
}

