% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgooglepubsub.R
\docType{package}
\name{rgooglepubsub}
\alias{rgooglepubsub}
\alias{rgooglepubsub-package}
\title{rgooglepubsub: Giving R users access to server utilities}
\description{
This package aims is to give R users the capability to have full programmatic
access to Google Pubsub tool
}

